Add docker deployment support (#268)
* try moving docker context to repo root

* docker.py cleanup

* Add docker deployment support

* Set up travis docker environment to test server docker build

* Replace django base image with python base image

because django dokcer image is deprecated

* change docker order

* debug

* Fix script name typo

* Add celery_docker, this could be better

*  In case our web container goes up faster than our DB, we should wait until
we can successfully connect.
This happens mostly during the first time we initialize docker-compose for the first run.
Originally, Docker Compose offers a healthcheck field in docker-compose.yml
however it doesn't seem to work properly and forces one to move to a newer docker-compose.yml vesrion (2.1)

Having tried the original solution of nc'ating to 3306 and waiting until it is open works.

p.s
this is exactly what the healthcheck field does, however for some reason it doesn't work properly,
may it be DNS, or some networking issues between the WEB container and the DB container.

* Run dockers in background and make tests run

* Try running tests inside docker

* Pass travis env to docker

* Avoid coveralls call in docker envs

* Remove coveralls installtion and usage